chore(broker-core): rename package from clientapi to commandapi
