+++ broker-core/src/main/java/io/zeebe/broker/transport/commandapi/CommandApiMessageHandlerService.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package io.zeebe.broker.transport.clientapi;
+package io.zeebe.broker.transport.commandapi;
 
 import io.zeebe.broker.clustering.base.partitions.Partition;
 import io.zeebe.servicecontainer.Service;
