+++ broker-core/src/test/java/io/zeebe/broker/engine/CreateDeploymentMultiplePartitionsTest.java
@@ -36,8 +36,8 @@ import io.zeebe.protocol.Protocol;
 import io.zeebe.protocol.clientapi.RecordType;
 import io.zeebe.protocol.clientapi.ValueType;
 import io.zeebe.protocol.intent.DeploymentIntent;
-import io.zeebe.test.broker.protocol.clientapi.CommandApiRule;
-import io.zeebe.test.broker.protocol.clientapi.ExecuteCommandResponse;
+import io.zeebe.test.broker.protocol.commandapi.CommandApiRule;
+import io.zeebe.test.broker.protocol.commandapi.ExecuteCommandResponse;
 import io.zeebe.test.util.record.RecordingExporter;
 import java.io.ByteArrayOutputStream;
 import java.nio.file.Files;
