+++ broker-core/src/test/java/io/zeebe/broker/engine/incident/ExpressionIncidentTest.java
@@ -35,8 +35,8 @@ import io.zeebe.protocol.ErrorType;
 import io.zeebe.protocol.clientapi.RecordType;
 import io.zeebe.protocol.intent.IncidentIntent;
 import io.zeebe.protocol.intent.WorkflowInstanceIntent;
-import io.zeebe.test.broker.protocol.clientapi.CommandApiRule;
-import io.zeebe.test.broker.protocol.clientapi.PartitionTestClient;
+import io.zeebe.test.broker.protocol.commandapi.CommandApiRule;
+import io.zeebe.test.broker.protocol.commandapi.PartitionTestClient;
 import io.zeebe.test.util.MsgPackUtil;
 import io.zeebe.test.util.collection.Maps;
 import java.util.List;
