+++ broker-core/src/test/java/io/zeebe/broker/engine/job/JobTimeOutTest.java
@@ -31,8 +31,8 @@ import io.zeebe.exporter.api.record.value.JobRecordValue;
 import io.zeebe.protocol.intent.Intent;
 import io.zeebe.protocol.intent.JobBatchIntent;
 import io.zeebe.protocol.intent.JobIntent;
-import io.zeebe.test.broker.protocol.clientapi.CommandApiRule;
-import io.zeebe.test.broker.protocol.clientapi.PartitionTestClient;
+import io.zeebe.test.broker.protocol.commandapi.CommandApiRule;
+import io.zeebe.test.broker.protocol.commandapi.PartitionTestClient;
 import java.time.Duration;
 import java.util.List;
 import java.util.stream.Collectors;
