+++ broker-core/src/test/java/io/zeebe/broker/exporter/ExporterManagerTest.java
@@ -29,8 +29,8 @@ import io.zeebe.model.bpmn.Bpmn;
 import io.zeebe.model.bpmn.BpmnModelInstance;
 import io.zeebe.msgpack.value.DocumentValue;
 import io.zeebe.protocol.intent.DeploymentIntent;
-import io.zeebe.test.broker.protocol.clientapi.CommandApiRule;
-import io.zeebe.test.broker.protocol.clientapi.PartitionTestClient;
+import io.zeebe.test.broker.protocol.commandapi.CommandApiRule;
+import io.zeebe.test.broker.protocol.commandapi.PartitionTestClient;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import org.junit.Before;
