+++ broker-core/src/test/java/io/zeebe/broker/system/UniqueKeyFormatTest.java
@@ -24,7 +24,7 @@ import io.zeebe.broker.test.EmbeddedBrokerRule;
 import io.zeebe.model.bpmn.Bpmn;
 import io.zeebe.protocol.Protocol;
 import io.zeebe.protocol.impl.record.value.workflowinstance.WorkflowInstanceCreationRecord;
-import io.zeebe.test.broker.protocol.clientapi.CommandApiRule;
+import io.zeebe.test.broker.protocol.commandapi.CommandApiRule;
 import io.zeebe.test.util.TestUtil;
 import io.zeebe.test.util.record.RecordingExporter;
 import org.junit.Rule;
