+++ protocol-test-util/src/main/java/io/zeebe/test/broker/protocol/commandapi/ErrorResponseException.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.zeebe.test.broker.protocol.clientapi;
+package io.zeebe.test.broker.protocol.commandapi;
 
 import io.zeebe.protocol.clientapi.ErrorCode;
